Add files via upload
I wanted to create a multi-card blackjack game to practice my python programming. The objective was to create a series of programs of limited functionality to work towards a final product in pieces. I started by creating a deck as a dictionary with a numerical index. I added parts piece by piece. Until I had a game without betting. 
Then I added betting and the ability to play the game more than once. I still have not implemented the possibility of splitting doubles. This is my next step. I also realized that there is probably a better way of handling the cards. I created a deck that is in order and then have cards drawn randomly from the deck while ensuring that the players don't draw the same card more than once. I realize now it would be better to shuffle the cards first. This would be more efficient as it would not require as much data to keep track of what cards have been played earlier